backend/requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
pytest==7.0.1
pytest-flask==1.3.0
flask_migrate==4.0.7
